requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.7.0; python_version >= '3.9'
asn1crypto==1.5.1
boto3==1.35.87; python_version >= '3.8'
botocore==1.35.87; python_version >= '3.8'
certifi==2024.12.14; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.8; python_version >= '3.7'
cloudpickle==2.2.1; python_version >= '3.6'
cryptography==44.0.0; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
fastapi==0.115.6; python_version >= '3.8'
filelock==3.16.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jmespath==1.0.1; python_version >= '3.7'
numpy==2.2.1; python_version == '3.11'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
platformdirs==4.3.6; python_version >= '3.8'
protobuf==5.29.2; python_version >= '3.8'
pyarrow==18.1.0
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.4; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pyopenssl==24.3.0; python_version >= '3.7'
pyral==1.6.0; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-http-client==3.3.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
redis==5.2.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
s3transfer==0.10.4; python_version >= '3.8'
sendgrid==6.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
setuptools==75.6.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
snowflake-connector-python[pandas]==3.12.4; python_version >= '3.8'
snowflake-snowpark-python[pandas]==1.26.0; python_version < '3.12' and python_version >= '3.8'
sortedcontainers==2.4.0
starkbank-ecdsa==2.2.0
starlette==0.41.3; python_version >= '3.8'
tomlkit==0.13.2; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
tzlocal==5.2; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.10'
uvicorn==0.34.0; python_version >= '3.9'
wheel==0.45.1; python_version >= '3.8'
